Update recipeListView when a recipe is updated
Fix the bug that when recipe is edited, it's updated in the database and map, however the listview is not updated.

Now the listview is connected to the observableList in the RecipeBooK class.
When opening the app, the list and the map in the Recipebook class get loaded from the database,
which separate the logic of saving and loading to database, while map and list works in memory with other operations when app is running.