Add delete function
User can now press delete button to toggle to delete mode, when the checkbox will show and the recipeView will be disabled.
User can select recipes to delete multiple ones at a time.
Also added that the search,add a new recipe and delete recipe are disabled when user click a recipe to view details.